[GEO] Refactor BaseGeoShapeFieldMapper to support Geometry Fields
This commit refactors BaseGeoShapeFieldMapper to a new BaseGeometryFieldMapper that
can be reused by other geometry field types. BaseGeoShapeFieldMapper is further refactored
to derive from the common logic provided by BaseGeometryFieldMapper needed to support
GeoShapeFieldMapper (LatLonShape types) and LegacyGeoShapeFieldMapper (prefix trees).
